requirements.txt
# core
cryptography
Pillow

# ui
pyperclip
urwid

# qa
flake8
pytest
black
